pyproject.toml
[tool.poetry]
name = "maestore"
version = "0.1.0"
description = ""
authors = ["pixDeVl <pixldevxp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
snowflake-tool = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
